confirmacion de inicializacion de la bd

+++ initDb.js
@@ -1,10 +1,24 @@
 "use strict";
 require("dotenv").config();
-const conn = require("./lib/connectMongoose");
+// const conn = require("./lib/connectMongoose");
 const fs = require("fs");
-
+const readline = require("readline");
 const sample = "./sample.json";
 const Item = require("./models/Item");
+const { resolve } = require("path");
+
+const confirm = function (question) {
+  return new Promise((resolve, reject) => {
+    const rl = readline.createInterface({
+      input: process.stdin,
+      output: process.stdout,
+    });
+    rl.question(question, function (answer) {
+      resolve(answer);
+      rl.close();
+    });
+  });
+};
 
 const loadJson = async function (file) {
   const dataPromise = new Promise((resolve, reject) => {
@@ -22,7 +36,7 @@ const loadJson = async function (file) {
     }
     const data = JSON.parse(rawData).items;
     console.log(`Initialization file ${file} parsed`);
-    console.log(data);
+    // console.log(data);
     return data;
   } catch (error) {
     console.log(error);
@@ -30,15 +44,24 @@ const loadJson = async function (file) {
 };
 
 const initialize = async function () {
-  try {
-    const result = await Item.deleteMany({});
-    console.log(`Deleted items: ${result.deletedCount}`);
-    const data = await loadJson(sample);
-    const newItems = await Item.insertMany(data);
-    console.log(`New inserted items: ${newItems.length}`);
-  } catch (error) {
-    console.log(error);
+  const answer = await confirm(
+    "Are you sure you want to initialize the database? (Yes)\n"
+  );
+  if (answer !== "Yes") {
+    process.exit(0);
+  } else {
+    try {
+      const conn = require("./lib/connectMongoose");
+      const result = await Item.deleteMany({});
+      console.log(`Deleted items: ${result.deletedCount}`);
+      const data = await loadJson(sample);
+      const newItems = await Item.insertMany(data);
+      console.log(`New inserted items: ${newItems.length}`);
+      conn.close();
+    } catch (error) {
+      console.log(error);
+    }
   }
 };
 
-initialize().then(() => conn.close());
+initialize();
